refactor(rich): consolidate rich logic into a more central place
