+++ ibis/expr/types/core.py
@@ -17,6 +17,7 @@ from ibis.common.typing import get_defining_scope
 from ibis.config import _default_backend
 from ibis.config import options as opts
 from ibis.expr.format import pretty
+from ibis.expr.types.rich import capture_rich_renderable, to_rich
 from ibis.util import experimental
 
 if TYPE_CHECKING:
@@ -27,41 +28,11 @@ if TYPE_CHECKING:
     import polars as pl
     import pyarrow as pa
     import torch
-    from rich.console import Console, RenderableType
+    from rich.console import Console
 
     import ibis.expr.types as ir
     from ibis.backends import BaseBackend
     from ibis.expr.visualize import EdgeAttributeGetter, NodeAttributeGetter
-
-
-try:
-    from rich.jupyter import JupyterMixin
-except ImportError:
-
-    class _FixedTextJupyterMixin:
-        """No-op when rich is not installed."""
-else:
-
-    class _FixedTextJupyterMixin(JupyterMixin):
-        """JupyterMixin adds a spurious newline to text, this fixes the issue."""
-
-        def _repr_mimebundle_(self, *args, **kwargs):
-            try:
-                bundle = super()._repr_mimebundle_(*args, **kwargs)
-            except Exception:  # noqa: BLE001
-                return None
-            else:
-                bundle["text/plain"] = bundle["text/plain"].rstrip()
-                return bundle
-
-
-def _capture_rich_renderable(renderable: RenderableType) -> str:
-    from rich.console import Console
-
-    console = Console(force_terminal=False)
-    with console.capture() as capture:
-        console.print(renderable)
-    return capture.get().rstrip()
 
 
 @public
@@ -80,7 +51,7 @@ class Expr(Immutable, Coercible):
 
     def __repr__(self) -> str:
         if ibis.options.interactive:
-            return _capture_rich_renderable(self)
+            return capture_rich_renderable(self)
         else:
             return self._noninteractive_repr()
 
@@ -101,8 +72,6 @@ class Expr(Immutable, Coercible):
 
         try:
             if opts.interactive:
-                from ibis.expr.types.pretty import to_rich
-
                 rich_object = to_rich(self, console_width=console_width)
             else:
                 rich_object = Text(self._noninteractive_repr())
